Created CurriculumServiceTest, added existsById verification on CurriculumService methods

+++ src/main/java/com/ifpb/lattesmaismais/business/CurriculumService.java
@@ -14,16 +14,22 @@ public class CurriculumService {
 	private CurriculumRepository curriculumRepository;
 	
 	public Curriculum findById(Integer id) throws ObjectNotFoundException {
-		return curriculumRepository.findById(id).orElseThrow(
-					() -> new ObjectNotFoundException(String.format("Currículo com id %d não encontrado!", id))
-				);
+		if (!curriculumRepository.existsById(id)) {
+			throw new ObjectNotFoundException(String.format("Currículo com id %d não encontrado!", id));
+		}
+
+		return curriculumRepository.findById(id).get();
 	}
 	
 	public Curriculum save(Curriculum curriculum) {
 		return curriculumRepository.save(curriculum);
 	}
 	
-	public void deleteById(Integer id) {
+	public void deleteById(Integer id) throws ObjectNotFoundException {
+		if (!curriculumRepository.existsById(id)) {
+			throw new ObjectNotFoundException(String.format("Currículo com id %d não encontrado!", id));
+		}
+
 		curriculumRepository.deleteById(id);
 	}
 }
